requirements.txt
rembg
reportlab
pillow
pandas
rarfile
patool
streamlit